chore(widget): solve dependency cycle
